requirements.txt
pylint==2.3.1
black==18.9b0
Django>=3.1.7
psycopg2==2.7.7
graphene-django>=2.0
django-graphql-jwt==0.2.0
django-jwt-auth==0.0.2
PyJWT==1.7.1
graphene-file-upload==1.2.2
flake8==3.7.7
Pillow==5.4.1
ipython===7.4.0
django-modeltranslation==0.16.2